fix type convertion

+++ lab2/task4/PrimeNumbers/PrimeNumbers.cpp
@@ -16,7 +16,7 @@ set<int> GeneratePrimeNumbersSet(int upperBound)
 {
 	vector<bool> sieve = GetSieve(upperBound);
 	set<int> primeNumbersSet;
-	for (size_t i = MIN_BOUND; i <= upperBound; i++)
+	for (int i = MIN_BOUND; i <= upperBound; i++)
 	{
 		if (sieve[i])
 		{
@@ -29,11 +29,11 @@ set<int> GeneratePrimeNumbersSet(int upperBound)
 vector<bool> GetSieve(int upperBound)
 {
 	vector<bool> sieve(upperBound + 1, true);
-	for (size_t i = MIN_BOUND; i * i <= upperBound; i++)
+	for (int i = MIN_BOUND; i * i <= upperBound; i++)
 	{
 		if (sieve[i])
 		{
-			for (size_t k = i * i; k <= upperBound; k += i)
+			for (int k = i * i; k <= upperBound; k += i)
 			{
 				sieve[k] = false;
 			}
